Rework partitioned data generation
There is currently no way to control the number of partitions,
as they are split based on number of lines. What can be controlled
the number of batches the tbl files are generated in, and the
paralellism used for each batch. The result is as many partitions
needed for each batch as there are rows per table.

Fixes:
- use of PartitionMaxSize
- bucket name when syncing to AWS S3